refactor: harden world map with static array and safe accessors

+++ src/constants.rs
@@ -1,4 +1,4 @@
-pub const MAP_WIDTH: i32 = 24;
-pub const MAP_HEIGHT: i32 = 24;
+pub const MAP_WIDTH: usize = 24;
+pub const MAP_HEIGHT: usize = 24;
 pub const SCREEN_WIDTH: u32 = 600;
 pub const SCREEN_HEIGHT: u32 = 400;
